Se agrego el metodo de busqueda por medio de modulo

+++ src/MetodosBusqueda/Lectura.java
@@ -34,7 +34,7 @@ public class Lectura{
         System.out.print(mens);
         cad=sc.nextLine();
         try{
-            x=Integer.parseInt(cad);
+            x=Long.parseLong(cad);
         }catch(NumberFormatException e){
             return EnteroL(mens);
         }
